Create enough space on LUNs now we're not double allocating(!)
The LUNs exposed were only 9.31GB (10000000000B) and we were creating a 10GB
VDI and filling it. This probably should have been an alarm bell that it was
deemed to be working.

However, now we've stopped double allocating blocks (by
https://github.com/xapi-project/xenvm/pull/136), we don't have enough space to
run the test (as we should expect).

This patch makes the LUNs just short of 20GB.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>